+++ jest.config.js
@@ -1,0 +1,3 @@
+export default {
+  testMatch: ['<rootDir>/src/**/*.test.js'],
+};