+++ src/lib/project.test.js
@@ -17,29 +17,66 @@ describe('project.js', () => {
   });
 
   describe('setProjectName()', () => {
-    it('README.md contains target text to replace', () => {
-      const readmeTs = fs.readFileSync(
-        path.join('templates', 'project-ts', 'README.md')
+    const DIR = 'temp-fixture-proj';
+    const NAME = 'my-cool-zkapp';
+    const TEMPLATE_DIR = 'templates/project-ts';
+
+    beforeAll(() => {
+      fs.copyFileSync(
+        path.join(TEMPLATE_DIR, 'README.md'),
+        path.join(TEMPLATE_DIR, 'README.md.bak')
       );
-      expect(readmeTs.includes('Mina zkApp: PROJECT_NAME')).toBeTruthy();
+      fs.copyFileSync(
+        path.join(TEMPLATE_DIR, 'package.json'),
+        path.join(TEMPLATE_DIR, 'package.json.bak')
+      );
+    });
+
+    afterAll(() => {
+      fs.copyFileSync(
+        path.join(TEMPLATE_DIR, 'README.md.bak'),
+        path.join(TEMPLATE_DIR, 'README.md')
+      );
+      fs.copyFileSync(
+        path.join(TEMPLATE_DIR, 'package.json.bak'),
+        path.join(TEMPLATE_DIR, 'package.json')
+      );
+      fs.unlinkSync(path.join(TEMPLATE_DIR, 'README.md.bak'));
+      fs.unlinkSync(path.join(TEMPLATE_DIR, 'package.json.bak'));
+    });
+
+    afterEach(() => {
+      if (fs.existsSync(DIR)) {
+        fs.rmSync(DIR, { recursive: true });
+      }
+    });
+
+    it('README.md contains target text to replace', () => {
+      const readme = fs.readFileSync(
+        path.join(TEMPLATE_DIR, 'README.md'),
+        'utf8'
+      );
+      expect(readme.includes('Mina zkApp: PROJECT_NAME')).toBeTruthy();
     });
 
     it('package.json contains target text to replace', () => {
-      const readmeTs = fs.readFileSync(
-        path.join('templates', 'project-ts', 'package.json')
+      const packageJson = fs.readFileSync(
+        path.join(TEMPLATE_DIR, 'package.json'),
+        'utf8'
       );
-      expect(readmeTs.includes('package-name')).toBeTruthy();
+      expect(packageJson.includes('package-name')).toBeTruthy();
     });
 
     it('should replace text in README.md & package.json', () => {
-      const DIR = 'temp-fixture-proj';
-      const NAME = 'my-cool-zkapp';
-
-      const README = '# Mina zkApp: PROJECT_NAME\n more stuff\n and more';
-      const PKG = `{"name": "package-name","version": "0.1.0"}`;
-      fs.mkdirSync('temp-fixture-proj', { recursive: true });
-      fs.writeFileSync(DIR + '/README.md', README);
-      fs.writeFileSync(DIR + '/package.json', PKG);
+      fs.mkdirSync(DIR, { recursive: true });
+      fs.writeFileSync(
+        DIR + '/README.md',
+        '# Mina zkApp: PROJECT_NAME\n more stuff\n and more'
+      );
+      fs.writeFileSync(
+        DIR + '/package.json',
+        `{"name": "package-name","version": "0.1.0"}`
+      );
 
       setProjectName(DIR, NAME);
 
@@ -52,20 +89,23 @@ describe('project.js', () => {
       );
       expect(packageAfter.includes('my-cool-zkapp')).toBeTruthy();
       expect(packageAfter.includes('package-name')).toBeFalsy();
-      fs.rmSync(DIR, { recursive: true });
     });
   });
 
   describe('replaceInFile()', () => {
+    const file = 'tmp-fixture-file';
+
+    afterEach(() => {
+      fs.unlinkSync(file);
+    });
+
     it('should replace target content in a file', () => {
-      const file = 'tmp-fixture-file';
       const str = '# Mina zkApp: PROJECT_NAME\n more stuff';
       fs.writeFileSync(file, str);
       replaceInFile(file, 'PROJECT_NAME', 'Foo Bar');
       const result = fs.readFileSync(file, 'utf8');
       expect(result.includes('Foo Bar')).toBeTruthy();
       expect(result.includes('PROJECT_NAME')).toBeFalsy();
-      fs.unlinkSync(file);
     });
   });